Fixed int vs unsigned

+++ connectiondialog.cpp
@@ -45,7 +45,7 @@ void ConnectionDialog::tableIfShowComputers(int id)
     ui->linkedTable->setRowCount((int)LinkedComputers.size());
     ui->linkedTable->clearContents();
     ui->linkedTable->setColumnCount(3);
-    for(int row = 0; row < LinkedComputers.size(); row++)
+    for(size_t row = 0; row < LinkedComputers.size(); row++)
     {
 
         Computer CurrentComputer = LinkedComputers.at(row);
@@ -70,7 +70,7 @@ void ConnectionDialog::tableIfShowPersons(int id)
 
     ui->linkedTable->setRowCount((int)LinkedPersons.size());
     ui->linkedTable->clearContents();
-    for(int row = 0; row < LinkedPersons.size(); row++)
+    for(size_t row = 0; row < LinkedPersons.size(); row++)
     {
         Person CurrentPerson = LinkedPersons.at(row);
         QString name = QString::fromStdString(CurrentPerson.getName());
@@ -97,9 +97,9 @@ void ConnectionDialog::unlinkTableIfShowComputers(int id)
     //UnLinkedComputers.clear();
     UnLinkedComputers =_CService.getVectorFromDataAccess(UnLinkedComputers);
 
-    for (int i = 0; i < LinkedComputers.size() ; i++)
+    for (size_t i = 0; i < LinkedComputers.size() ; i++)
     {
-        for(int j = 0; j < UnLinkedComputers.size() ; j++)
+        for(size_t j = 0; j < UnLinkedComputers.size() ; j++)
         {
             if( ( LinkedComputers.at(i).getName() == UnLinkedComputers.at(j).getName() ) &
                  ( LinkedComputers.at(i).getType() == UnLinkedComputers.at(j).getType() )  &
@@ -113,7 +113,7 @@ void ConnectionDialog::unlinkTableIfShowComputers(int id)
     ui->unLinkedTable->setRowCount((int)UnLinkedComputers.size());
     ui->unLinkedTable->clearContents();
     ui->unLinkedTable->setColumnCount(3);
-    for(int row = 0; row < UnLinkedComputers.size(); row++)
+    for(size_t row = 0; row < UnLinkedComputers.size(); row++)
     {
 
         Computer CurrentComputer = UnLinkedComputers.at(row);
@@ -138,9 +138,9 @@ void ConnectionDialog::unlinkTableIfShowPersons(int id)
     UnLinkedPersons=_PService.getVectorFromDataAccess(UnLinkedPersons);
 
     qDebug() << "e";
-    for (int i = 0; i < LinkedPersons.size() ; i++)
+    for (size_t i = 0; i < LinkedPersons.size() ; i++)
     {
-        for(int j = 0; j < UnLinkedPersons.size() ; j++)
+        for(size_t j = 0; j < UnLinkedPersons.size() ; j++)
         {
             if( ( LinkedPersons.at(i).getName() == UnLinkedPersons.at(j).getName() ) &
                 ( LinkedPersons.at(i).getGender() == UnLinkedPersons.at(j).getGender() )  &
@@ -154,7 +154,7 @@ void ConnectionDialog::unlinkTableIfShowPersons(int id)
 
     ui->unLinkedTable->setRowCount((int)UnLinkedPersons.size());
     ui->unLinkedTable->clearContents();
-    for(int row = 0; row < UnLinkedPersons.size(); row++)
+    for(size_t row = 0; row < UnLinkedPersons.size(); row++)
     {
         Person CurrentPerson = UnLinkedPersons.at(row);
         QString name = QString::fromStdString(CurrentPerson.getName());
